requirements.txt
discord.py
requests
apscheduler
pytz
